[REFACTOR] 컴포넌트 index.js 정의
