+++ components/plan/btn/index.js
@@ -1,0 +1,3 @@
+import OppositeButton from './OppositeButton';
+
+export { OppositeButton };
