+++ components/plan/modal/OppositeModal.jsx
