+++ components/plan/modal/index.js
@@ -1,0 +1,3 @@
+import OppositeModal from './OppositeModal';
+
+export { OppositeModal };
